Extract RemaingTaskRow in its own component
