+++ tasks-app-shared/src/commonTest/kotlin/net/opatry/tasks/ui/component/DateRangeColorTest.kt
@@ -29,8 +29,8 @@ import androidx.compose.ui.test.runComposeUiTest
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.minus
 import kotlinx.datetime.plus
+import net.opatry.tasks.app.ui.component.toColor
 import net.opatry.tasks.app.ui.model.DateRange
-import net.opatry.tasks.app.ui.screen.toColor
 import net.opatry.tasks.ui.screen.today
 import kotlin.test.Test
 import kotlin.test.assertEquals
