+++ tasks-app-shared/src/commonTest/kotlin/net/opatry/tasks/ui/component/RemainingTaskRowTest.kt
@@ -20,7 +20,7 @@
  * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.opatry.tasks.ui.screen
+package net.opatry.tasks.ui.component
 
 import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.assertCountEquals
@@ -38,6 +38,7 @@ import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.runComposeUiTest
 import kotlinx.datetime.DateTimeUnit
 import kotlinx.datetime.minus
+import net.opatry.tasks.app.ui.component.RemainingTaskRow
 import net.opatry.tasks.app.ui.component.TaskAction
 import net.opatry.tasks.app.ui.component.TaskMenuTestTag.ADD_SUBTASK
 import net.opatry.tasks.app.ui.component.TaskMenuTestTag.DELETE
@@ -47,7 +48,6 @@ import net.opatry.tasks.app.ui.component.TaskMenuTestTag.MOVE_TO_NEW_LIST
 import net.opatry.tasks.app.ui.component.TaskMenuTestTag.MOVE_TO_TOP
 import net.opatry.tasks.app.ui.component.TaskMenuTestTag.TASK_MENU
 import net.opatry.tasks.app.ui.component.TaskMenuTestTag.UNINDENT
-import net.opatry.tasks.app.ui.screen.RemainingTaskRow
 import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.REMAINING_TASK_DUE_DATE_CHIP
 import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.REMAINING_TASK_ICON
 import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.REMAINING_TASK_MENU_ICON
@@ -55,6 +55,9 @@ import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.REMAINING_TASK_NOTES
 import net.opatry.tasks.app.ui.screen.TaskListPaneTestTag.REMAINING_TASK_ROW
 import net.opatry.tasks.resources.Res
 import net.opatry.tasks.resources.task_due_date_label_weeks_ago
+import net.opatry.tasks.ui.screen.createTask
+import net.opatry.tasks.ui.screen.createTaskList
+import net.opatry.tasks.ui.screen.today
 import org.jetbrains.compose.resources.pluralStringResource
 import kotlin.test.Ignore
 import kotlin.test.Test
